download route alteration
