+++ src/app/cube/details/components/action-panel/action-panel.component.ts
@@ -22,6 +22,8 @@ import { LearningObjectService } from '../../../../../app/onion/core/learning-ob
 import { NavbarDropdownService } from '../../../../core/client-module/navBarDropdown.service';
 import { BUNDLING_ROUTES } from 'app/core/learning-object-module/bundling/bundling.routes';
 import { HttpErrorResponse } from '@angular/common/http';
+import { BundlingService } from 'app/core/learning-object-module/bundling/bundling.service';
+
 
 @Component({
   selector: 'cube-details-action-panel',
@@ -83,7 +85,8 @@ export class ActionPanelComponent implements OnInit, OnDestroy {
     private collectionService: CollectionService,
     private router: Router,
     private learningObjectService: LearningObjectService,
-    private dropdowns: NavbarDropdownService
+    private dropdowns: NavbarDropdownService,
+    private bundlingService: BundlingService
   ) { }
 
   async ngOnInit(): Promise<void> {
@@ -113,7 +116,7 @@ export class ActionPanelComponent implements OnInit, OnDestroy {
 
   get mapAndTag() {
     const untaggable = this.learningObject.status === LearningObject.Status.RELEASED
-                              || this.learningObject.status === LearningObject.Status.UNRELEASED;
+      || this.learningObject.status === LearningObject.Status.UNRELEASED;
     if (this.auth.user && this.auth.user.accessGroups && !this.userIsAuthor && !untaggable) {
       const privileges = ['admin', 'editor', 'mapper'];
       if (this.auth.user.accessGroups.some(priv => privileges.includes(priv))) {
@@ -139,7 +142,7 @@ export class ActionPanelComponent implements OnInit, OnDestroy {
       this.addingToLibrary = true;
     }
     if (download) {
-      this.download(
+      this.bundlingService.download(
         this.learningObject.id
       );
     }
@@ -187,7 +190,7 @@ export class ActionPanelComponent implements OnInit, OnDestroy {
    */
   downloadRevised(download?: boolean) {
     if (download) {
-      this.download(
+      this.bundlingService.download(
         this.learningObject.id
       );
     }
@@ -199,10 +202,10 @@ export class ActionPanelComponent implements OnInit, OnDestroy {
    * @param learningObjectId the unique mongo id of a learning object
    */
 
-  download(learningObjectId: string) {
-    this.toggleDownloadModal(true);
-    this.libraryService.downloadBundle(BUNDLING_ROUTES.DOWNLOAD_BUNDLE(learningObjectId));
-  }
+  // download(learningObjectId: string) {
+  //   this.toggleDownloadModal(true);
+  //   this.bundlingService.downloadBundle(BUNDLING_ROUTES.DOWNLOAD_BUNDLE(learningObjectId));
+  // }
 
   copyLink() {
     const el = this.objectLinkElement.nativeElement;
